Move test method
